Update main.c
myRoot の y は double 型に
それに伴い myRoot 自体の戻り値の定義も double 型に

+++ main.c
@@ -2,9 +2,9 @@
 #include <stdlib.h>
 #include <math.h>
 
-int myRoot(double x) // x の平方根を返す関数。x >= 0 ではないとき exit(1) とする。 
+double myRoot(double x) // x の平方根を返す関数。x >= 0 ではないとき exit(1) とする。 
 {
-  int y;
+  double y;
   if( x <= 0 ){
     exit(1);
   }
